RoadDamageDetection-main/requirements.txt
numpy==1.24.3
Pillow==9.5.0
streamlit==1.28.0
torch==2.0.0
torchvision==0.15.1
ultralytics>=8.2.0
opencv-python-headless==4.8.0.74
requests==2.28.2
av>=11.0.0
streamlit-webrtc>=0.47.1
wheel
